move runways lookup into its own utils file
